+++ examples/server/main.go
@@ -100,15 +100,24 @@ func segments(w http.ResponseWriter, r *http.Request) {
 
 func mysql(w http.ResponseWriter, r *http.Request) {
 	txn, _ := w.(newrelic.Transaction)
-	defer newrelic.DatastoreSegment{
-		StartTime:  newrelic.StartSegmentNow(txn),
-		Product:    newrelic.DatastoreMySQL,
-		Collection: "my_table",
-		Operation:  "SELECT",
-	}.End()
+	s := newrelic.DatastoreSegment{
+		StartTime:          newrelic.StartSegmentNow(txn),
+		Product:            newrelic.DatastoreMySQL,
+		Collection:         "users",
+		Operation:          "INSERT",
+		ParameterizedQuery: "INSERT INTO users (name, age) VALUES ($1, $2)",
+		QueryParameters: map[string]interface{}{
+			"name": "Dracula",
+			"age":  439,
+		},
+		Host:         "mysql-server-1",
+		PortPathOrID: "3306",
+		DatabaseName: "my_database",
+	}
+	defer s.End()
 
 	time.Sleep(20 * time.Millisecond)
-	io.WriteString(w, `performing fake query "SELECT * from my_table"`)
+	io.WriteString(w, `performing fake query "INSERT * from users"`)
 }
 
 func external(w http.ResponseWriter, r *http.Request) {